requirements_complete_verifier.txt
appdirs==1.4.4
coloredlogs==15.0.1
filelock==3.15.1
flatbuffers==24.3.25
fsspec==2024.2.0
humanfriendly==10.0
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnx2pytorch @ git+https://github.com/KaidiXu/onnx2pytorch@fe7281b9b6c8c28f61e72b8f3b0e3181067c7399
onnxconverter-common==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.15.1
onnxsim==0.4.35
packaging==24.1
pandas==2.1.4
pillow==10.3.0
protobuf==3.20.2
psutil==5.9.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
skl2onnx==1.16.0
sortedcontainers==2.4.0
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.65.0
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
